requirements.txt
watchdog
yara-python
flask
python-dotenv
python-magic
pillow
python-magic==0.4.27
#yara-python==4.2.3
watchdog==3.0.0
flask==3.0.0
python-dotenv==1.0.0
